requirements.txt
PyYAML==6.0
kafka-python==2.0.2
psycopg2-binary==2.9.5
pytest==7.2.0
requests==2.28.1